requirements.inference.txt
# Минимальные зависимости для инференса моделей в handwritten-ocr
# (скрипты src/infer_page.py и src/eval_folder.py)

torch>=2.0.0
Pillow>=9.0.0
numpy>=1.21.0

# TrOCR
transformers>=4.30.0
sentencepiece>=0.1.99

# Детектор строк (PaddleOCR) для нарезки страниц
paddleocr>=2.7.0
paddlepaddle>=2.5.0

# Метрики (CER/WER)
python-Levenshtein>=0.23.0

